Introduce `ApiService` and `OtpVerificationRequest` for handling user authentication, menu, and order network requests.

+++ CollegeCanteen/app/src/main/java/com/collegecanteen/network/ApiService.java
@@ -35,5 +35,5 @@ public interface ApiService {
 
     @retrofit2.http.FormUrlEncoded
     @POST("/api/orders/update-status")
-    Call<ApiResponse> updateOrderStatus(@retrofit2.http.Field("orderId") int orderId, @retrofit2.http.Field("status") String status);
+    Call<ApiResponse> updateOrderStatus(@retrofit2.http.Field("orderId") Long orderId, @retrofit2.http.Field("status") String status);
 }
